cleaned up unused components and code snippets
