--- code/src/components/Bunny.js
@@ -1,11 +1,0 @@
-
-import React from 'react';
-import Spline from '@splinetool/react-spline';
-
-export default function App() {
-  return (
-    <div>
-	  <Spline scene="https://draft.spline.design/QViAyLB2id0atWsz/scene.spline" />
-	</div>
-  );
-}
